% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qf_flags.R
\name{check_qf_flags}
\alias{check_qf_flags}
\title{Internal helper function to determine availability of data within a time interval}
\usage{
check_qf_flags(measurement_name, data)
}
\arguments{
\item{measurement_name}{The name of the measurement (staPres, soilTemp, VSWC, soilCO2Concentration)}

\item{data}{Data used to check the qf flags}
}
\value{
A data frame of startDateTime, horizontalPosition, and the associated QF flag.

# changelog and author contributions / copyrights
}
\description{
Given a data measurement, determine when a monthly mean is used and the bulk QF flags. Helps to determine when the environmental measurements produced a QF value and to be used in subsequent flux calculations.
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
