% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_monthly_mean.R
\name{compute_monthly_mean}
\alias{compute_monthly_mean}
\title{Function to compute monthly means to smooth the data.}
\usage{
compute_monthly_mean(
  NEON_data,
  position_columns = c("horizontalPosition", "verticalPosition")
)
}
\arguments{
\item{NEON_data}{Required. Input vector of neon measurements for a month}

\item{position_columns}{Optional. Do we group by horiztonalPosition, verticalPosition, and? Default is both. Added this option in case we just want to average across a given dimension.}
}
\value{
A data frame that reports for each horiztonal and vertical position the computed mean and standard deviation from sampling (similar to a bootstrap method) as well as the sample mean and sample standard deviation
}
\description{
Given a NEON measurement data frame calculate the monthly mean values across all horizontal and vertical locations. Based off code from Zoey Werbin.
}
\examples{
# Download the NEON data - here this would be soil moisture
NEON_moist_30m_orig <-  neonUtilities::loadByProduct(dpID="DP1.00094.001",
site="WREF",
startdate="2022-06",
enddate="2022-06",
package="expanded",
check.size = F)

# Then correct the swc
site_swc <- swc_correct(NEON_moist_30m_orig,"WREF")

# Select the columns and the time frequency
time_frequency = "30_minute"
column_selectors = c("Mean","Minimum","Maximum","ExpUncert","StdErMean")

swc <- site_swc |>
pluck(paste0("SWS_",time_frequency)) |>
select(domainID,siteID,horizontalPosition,verticalPosition,startDateTime,matches(str_c("VSWC",column_selectors)),VSWCFinalQF)

# Now apply the monthly mean:
swc_monthly_mean <- compute_monthly_mean(swc)
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
