% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_surface_flux.R
\name{compute_surface_flux}
\alias{compute_surface_flux}
\title{Compute CO2 surface flux}
\usage{
compute_surface_flux(input_data)
}
\arguments{
\item{input_data}{Required. A data frame containing zOffsets, diffusivity, and co2 (umol mol-1)}
}
\value{
A value of the surface CO2 flux (umol m-2 s-1)
}
\description{
Computation function. Given a measurement of the co2 and diffusive flux at different levels, return the surface flux
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
