% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_surface_flux_layer.R
\name{compute_surface_flux_layer}
\alias{compute_surface_flux_layer}
\title{Title Determine the surface flux from linear regression at layers}
\usage{
compute_surface_flux_layer(input_data)
}
\arguments{
\item{input_data}{Required. A data frame containing zOffsets, diffusivity, and co2 (umol mol-1) and their associated errors}
}
\value{
A value of the surface CO2 flux (umol m-2 s-1)
}
\description{
Computation function. Given a measurement of the co2 and diffusive flux at different levels, return the surface flux
}
\seealso{
[compute_surface_flux()] for other ways to compute surface fluxes.
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
