% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_interpolate.R
\name{depth_interpolate}
\alias{depth_interpolate}
\title{Interpolate different depth measurements}
\usage{
depth_interpolate(
  input_measurements,
  measurement_name,
  measurement_interpolate,
  column_selectors
)
}
\arguments{
\item{input_measurements}{Required. Nested data frame (tibble of a months worth of data of co2, temperature, swc, pressure)}

\item{measurement_name}{Required. Names of measurements we are interpolating. Currently only does one column at a time.}

\item{measurement_interpolate}{Required. Names of measurement whose depth is used to interpolate (typically co2)}

\item{column_selectors}{Required. Names of columns (e.g column_selectors = c("Mean","ExpUncert")) that are used in the interpolation. These are matched for each NEON measurement, i.e SWC is VSWCMean for the Mean SWC.}
}
\value{
A nested data frame with interpolated measurements.
}
\description{
Definition function. Linearly interpolate a measurement across the different measurement depths
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
based on code developed by Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
