% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_position.R
\name{determine_position}
\alias{determine_position}
\title{Determine the depth of a measurement}
\usage{
determine_position(input_positions, input_measurement)
}
\arguments{
\item{input_positions}{Required. Input vector of measurements depths}

\item{input_measurement}{Required. Input vector of measurements.}
}
\value{
A data frame that reports the measurement depth for and associated environmental measurement.
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\description{
Given a NEON measurement data frame and measurement depth, pull out the measurement depth for a measurement - because it may vary in time based on if a sensor is replaced.
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
