% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env_data.R
\name{extract_env_data}
\alias{extract_env_data}
\title{Return NEON environmental data}
\usage{
extract_env_data(input_file_name, save_file = FALSE, output_file_name = NULL)
}
\arguments{
\item{input_file_name}{Required. Path of location for the save file from acquire_neon_data. Must end in .Rda (a string)}

\item{save_file}{Optional. Logical. Do we save the file or return it to console?}

\item{output_file_name}{Optional. Path of location for file to save when save_file is TRUE. Must end in .Rda (a string)}
}
\value{
Data frame of fluxes and gradient from the timeperiod
}
\description{
Given a site filename (from acquire_neon_data), process and compute fluxes.
This file takes a saved data file from acquire:
1) Takes the needed components (QF and measurement flags) for soil water, temperature, co2, binding them together in a tidy data frame
2) Interpolates across the measurements
3) Merges air pressure data into this data frame
4) Does a final QF check so we should have only timeperiods where all measurements exist
5) Adds in the megapit data so we have bulk density, porosity measurements at the interpolated depth.
6) Saves the data
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
