% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_mean.R
\name{insert_mean}
\alias{insert_mean}
\title{Internal function that inserts smoothed mean value of a measurement at a site}
\usage{
insert_mean(data, monthly_mean, measurement_name)
}
\arguments{
\item{data}{Required. input data to use}

\item{monthly_mean}{Required. monthly mean of input data to use}

\item{measurement_name}{Required. name of measurement}
}
\value{
Nested data frame of measurements
}
\description{
Given a site measurement and monthly mean, insert in the monthly mean value when the QF flag fails.
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
