% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_api_token.R
\name{neon_api_token}
\alias{neon_api_token}
\title{Install a NEON API token in Your \code{.Renviron} File for repeated Use}
\usage{
neon_api_token(token, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{token}{The API token provided to you from the NEON formatted in quotes. A token can be acquired at \url{https://www.neonscience.org/resources/learning-hub/tutorials/api-token-setup}.}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing NEON_TOKEN that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the token in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\description{
This function will add your NEON API token to your \code{.Renviron} file so it can be called securely without being stored in your code. After you have installed your key, it can be called any time by typing \code{Sys.getenv("NEON_TOKEN")} and can be used in package functions by simply typing NEON_TOKEN If you do not have an \code{.Renviron} file, the function will create on for you.
If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your original file for disaster recovery purposes.
This function was adapted from \code{census_api_key()} in the \code{tidycensus} library. \url{https://cran.r-project.org/web/packages/tidycensus/index.html}.
}
\examples{

\dontrun{
neon_api_token("111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("NEON_API_TOKEN")
}

\dontrun{
# If you need to overwrite an existing key:
neon_api_token("111111abc", overwrite = TRUE, install = TRUE)
# First time, relead your environment so you can use the token without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("NEON_TOKEN")
}
}
\seealso{
[get_neon_api_token()]
}
