% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_site_flux.R
\name{neon_site_flux}
\alias{neon_site_flux}
\title{Compute CO2 surface flux at a given NEON site}
\usage{
neon_site_flux(site_data, positions)
}
\arguments{
\item{site_data}{Required. A data frame containing zOffsets, positions, megapit data, depths, measurement data (computed in interpolate}
}
\value{
A value of the surface CO2 flux (umol m-2 s-1)
}
\description{
Computation function. Given a measurement of the co2 and diffusive flux at different levels, return the surface flux
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
