% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEONSoils-package.R
\docType{package}
\name{neonSoilFlux-package}
\alias{neonSoilFlux}
\alias{neonSoilFlux-package}
\title{neonSoilFlux: Compute Soil Carbon Fluxes for the National Ecological Observatory Network Sites}
\description{
Acquires and synthesizes soil carbon fluxes at sites located in the National Ecological Observatory Network (NEON). Provides flux estimates and associated uncertainty as well as key environmental measurements (soil water, temperature, CO2 concentration) that are used to compute soil fluxes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmzobitz/neonSoilFlux}
  \item Report bugs at \url{https://github.com/jmzobitz/neonSoilFlux/issues}
}

}
\author{
\strong{Maintainer}: John Zobitz \email{zobitz@augsburg.edu} (\href{https://orcid.org/0000-0002-1830-143X}{ORCID})

Authors:
\itemize{
  \item Edward Ayres \email{eayres@battelleecology.org} (\href{https://orcid.org/0000-0003-1846-1473}{ORCID})
}

Other contributors:
\itemize{
  \item Katie O'Rourke (LinkedIn, <https://www.linkedin.com/in/katie-o-rourke-5a898b194/>) [contributor]
  \item Zoey Werbin \email{zrwerbin@bu.edu} [contributor]
  \item Lajntxiag Lee \email{lajntxiag@alumni.augsburg.edu} [contributor]
  \item Ridwan Abdi (LinkedIn, <https://www.linkedin.com/in/ridwaan-cabdi/>) [contributor]
  \item Dijonë Mehmeti (LinkedIn, <https://www.linkedin.com/in/dijon%C3%AB-mehmeti/>) [contributor]
  \item Ly Xiong (LinkedIn, <https://www.linkedin.com/in/ly-xiong-bb83ba1b2/>) [contributor]
}

}
\keyword{internal}
