% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\name{powerset}
\alias{powerset}
\title{Compute all possible combinations of a vector - ignoring the empty set}
\usage{
powerset(x, size_start = 1)
}
\arguments{
\item{x}{Required. Input vector}

\item{size_start}{First power set input size}
}
\value{
A list of possible options
}
\description{
Compute all possible combinations of a vector - ignoring the empty set
}
\examples{
# All possible combinations
powerset(1:4)

# Now try without ignoring the single sets:
powerset(1:4,2)
}
