% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_error.R
\name{quadrature_error}
\alias{quadrature_error}
\title{Helper function to quickly compute the quadrature error}
\usage{
quadrature_error(x, x_err)
}
\arguments{
\item{x}{Required. Input vector of partial derivatives}

\item{x_err}{Required. Error vector of measurements}
}
\value{
A value of quadrature error
}
\description{
Given a vector (x) and its errors (x_err), compute the quadrature error
}
\examples{
TBD

}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
