% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_positions.R
\name{sensor_positions}
\alias{sensor_positions}
\title{Save the location of sensor positions for a given data product}
\usage{
sensor_positions(neon_data, data_code, data_product_id, save_filename)
}
\arguments{
\item{neon_data}{Required. A list of NEON data downloaded from the utilities}

\item{data_code}{Required. Names of data product we are interpolating. (SWS = soil water, ST = soil temperature, SCO2C = soil CO2)}

\item{data_product_id}{Name of the data product 00094 = soil water, 00041 = soil temperature, 00095 = soil CO2}
}
\value{
A data frame of the requested data.
}
\description{
Definition function. Given a measurement return the necessary components we need to compute flux calculations (and nothing more). Currently only works with 30 minute data
}
\examples{
sensor_positions(sjer_co2,"SCO2C","00095")
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{

}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
