% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjer_megapit_data_2022_06.R
\docType{data}
\name{sjer_megapit_data_2022_06}
\alias{sjer_megapit_data_2022_06}
\title{Measured soil physical properties at a NEON site}
\format{
A nested 10 item list
}
\source{
From \url{https://data.neonscience.org/data-products/DP1.00096.001}
}
\usage{
data(sjer_megapit_data_2022_06)
}
\description{
A nested dataset containing soil data properties at the SJER site from June 2022. A convenience data frame to make processing fluxes easier for testing.  Computed with using the function acquire_neon_data. Essentially what is returned when data product DP1.00096.001 is loaded from \link[neonUtilities]{loadByProduct}.
}
\details{
\itemize{
  \item citation_00096_RELEASE-2024: BibTex reference
  \item issueLog_00096: Listing of known issues
  \item mgp_perarchivesample: Archived sample information
  \item mgp_perbiogeosample: Data collected on biogeochemistry sample
  \item mgp_perbulksample: Data collected on bulk density sample
  \item mgp_perhorizon: Per soil horizon metadata
  \item mgp_permegapit: Data collected per megapit
  \item readme_00096: Data product description, issue log, and other metadata about the data product
  \item validation_00096: Description of data validation applied at the points of collection and ingest
  \item variables_00096: Description and units for each column of data in data tables
}
}
\keyword{datasets}
