% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tang_2003_flux.R
\name{tang_2003_flux}
\alias{tang_2003_flux}
\title{Compute surface flux at a given timepoint using a surface gradient via Tang et al 2003}
\usage{
tang_2003_flux(zOffset, co2, co2_err, diffusive, diffusive_err)
}
\arguments{
\item{zOffset}{Required. depths below surface - assumed to be positive in value. Important for directionality!}

\item{co2}{Required. co2 at depth (umol m--1)}

\item{co2_err}{Required. Associated errors with that value of co2}

\item{diffusive}{Required. diffusivity at each depth}

\item{diffusive_err}{Required Associated errors with diffusivity}
}
\value{
Data frame of fluxes associated error
}
\description{
Given  zOffsets, diffusivity, and co2 (umol mol-1) and their associated errors, compute the surface flux. This is done by computing the gradient of co2 (slope from linear regression) and linear extrapolation of surface diffusivity. Modified from Tang et al (2003).
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
