% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_neon_data.R
\name{acquire_neon_data}
\alias{acquire_neon_data}
\title{Acquire NEON data for processing}
\usage{
acquire_neon_data(
  site_name,
  start_date,
  end_date,
  file_name,
  env_file_name = NULL
)
}
\arguments{
\item{site_name}{Required. NEON code for a particular site (a string)}

\item{start_date}{Required. Date where we end getting NEON data. Format: YYYY-MM (can't specify day).  So "2020-05" means it will grab data for the entire 5th month of 2020. (a string)}

\item{end_date}{Required. Date where we end getting NEON data. Format: YYYY-MM (can't specify day).  So "2020-08" means it will grab data for the entire 8th month of 2020. (a string)}

\item{file_name}{Required. Path of location for the save file. Must end in .Rda (a string)}
}
\value{
Nothing is returned - the file is saved to the location provided
}
\description{
Given a site code and dates, apply the neonUtilities package to download the data from NEON API
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
