% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_to_umol.R
\name{co2_to_umol}
\alias{co2_to_umol}
\title{Convert co2 concentration from ppm to µmol m-3 units}
\usage{
co2_to_umol(temperature, pressure, co2)
}
\arguments{
\item{temperature}{Required. Soil temperature (degrees C)}

\item{pressure}{Required. Barometric air pressure ?}

\item{co2}{Carbon dioxide in ppm}
}
\value{
A value of the converted co2
}
\description{
Given a measurement of co2, convert it from ppm to umol m-3 based on temperature and pressure
}
\examples{
TBD

}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{

}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
