% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejong_shappert_flux.R
\name{dejong_shappert_flux}
\alias{dejong_shappert_flux}
\title{Compute surface flux at a given timepoint using a surface gradient via De Jong and Schappert (1972)}
\usage{
dejong_shappert_flux(zOffset, co2, co2_err, diffusive, diffusive_err)
}
\arguments{
\item{zOffset}{Required. depths below surface - assumed to be positive in value. Important for directionality!}

\item{co2}{Required. co2 at depth (umol m--1)}

\item{co2_err}{Required. Associated errors with that value of co2}

\item{diffusive}{Required. diffusivity at each depth}

\item{diffusive_err}{Required Associated errors with diffusivity}
}
\value{
Data frame of fluxes associated error
}
\description{
Given  zOffsets, diffusivity, and co2 (umol mol-1) and their associated errors, compute the surface flux. This is done by estimating the surface concentration (through linear regression) and doing a gradient calculation. Modified from De Jong and Schappert (1972).
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
