% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_fingerprint_plot.R
\name{env_fingerprint_plot}
\alias{env_fingerprint_plot}
\title{Helper function to plot QF results for environmental measurements.}
\usage{
env_fingerprint_plot(input_fluxes)
}
\arguments{
\item{input_fluxes}{data frame of computed fluxes}
}
\value{
A ggplot graph where we have ordered factors showing the QA values a given environmental measurement
}
\description{
Given a flux measurement data frame, show when the environmental measurements produced a QF value
}
\examples{
# Say you have a file name of computed fluxes:
load("my-fluxes.Rda") # Loads up out_fluxes
env_fingerprint_plot(out_fluxes)
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
