% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_function.R
\name{fit_function}
\alias{fit_function}
\title{Interpolate a soil measurement to different depths}
\usage{
fit_function(input_data, indep_name, dep_name, interp_values)
}
\arguments{
\item{input_data}{Required. Vector of NEON data and measurements from the soil}

\item{col_name}{Required. Names of column we are interpolating. Currently only does one column at a time.}
}
\value{
A data frame of the depth and the measured column
}
\description{
Definition function. Linearly interpolate a measurement across the different measurement depths
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{

}
\author{
John Zobitz \email{zobitz@augsburg.edu}
based on code developed by Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
