% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_api_token.R
\name{get_neon_api_token}
\alias{get_neon_api_token}
\title{Check to see if a NEON api token is installed}
\usage{
get_neon_api_token(token)
}
\arguments{
\item{key}{The API token provided to you from the Census formated in quotes. A key can be acquired at \url{https://www.neonscience.org/resources/learning-hub/tutorials/api-token-setup}}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing NEON_API_TOKEN that you already have in your \code{.Renviron} file.}
}
\description{
This function check in a NEON API token exists in your \code{.Renviron} file so it can be called securely without being stored in your code.
}
\seealso{
[neon_api_token()]
}
