% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_interval.R
\name{inside_interval}
\alias{inside_interval}
\title{Determine if a YYYY-MM string is inside a interval}
\usage{
inside_interval(start, end, reference_time)
}
\arguments{
\item{start}{starting interval time}

\item{end}{ending interval time}

\item{reference_time}{time we are comparing to - YYYY-MM string}
}
\value{
Logical indicating whether or not the reference time is inside the interval
}
\description{
Determine if a YYYY-MM string is inside a interval
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
