% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_merge.R
\name{measurement_merge}
\alias{measurement_merge}
\title{Acquire data from a NEON site}
\usage{
measurement_merge(
  neon_data,
  data_code,
  data_product_id,
  measurement_name,
  qf_name
)
}
\arguments{
\item{neon_data}{Required. A list of NEON data downloaded from the utilities}

\item{data_code}{Required. Names of data product we are interpolating. (SWS = soil water, ST = soil temperature, SCO2C = soil CO2)}

\item{data_product_id}{Name of the data product 00094 = soil water, 00041 = soil temperature, 00095 = soil CO2}

\item{measurement_name}{Required. Names of column we are grabbing. (VSWCMean = soil water, soilTempMean = soil temperature, soilCO2concentrationMean = soil CO2)}

\item{qf_name}{Required. Names of qf column we are grabbing. (VSWCFinalQF = soil water, finalQF = soil temperature, finalQF = soil CO2)}
}
\value{
A data frame of the requested data.
}
\description{
Definition function. Given a measurement return the necessary components we need to compute flux calculations (and nothing more). Currently only works with 30 minute data
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{

}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
