% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEONSoils-package.R
\docType{package}
\name{NEONSoils-package}
\alias{NEONSoils}
\alias{NEONSoils-package}
\title{NEONSoils: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
