% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_correct.R
\name{swc_correct}
\alias{swc_correct}
\title{Convert co2 concentration from ppm to µmol m-3 units}
\usage{
swc_correct(input_swc, curr_site)
}
\arguments{
\item{input_swc}{Required. input soil water content data from acquire_neon_data (as a list)}

\item{curr_site}{Current site we are working with}
}
\value{
A revised list
}
\description{
Given the expanded SWC data, return a corrected version based on the values below
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
\keyword{Currently}
\keyword{none}
