% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_correct.R
\name{swc_correct}
\alias{swc_correct}
\title{Internal function to correct depths for VSWC NEON data.}
\usage{
swc_correct(input_swc, curr_site, reference_time)
}
\arguments{
\item{input_swc}{Required. input soil water content data from acquire_neon_data (as a list)}

\item{curr_site}{Current site we are working with}

\item{reference_time}{Current month we are working with}
}
\value{
A revised list of corrected soil water content and depths.
}
\description{
Given the expanded SWC data, return a corrected version based on the values below
}
\details{
# (From correspondence w/ Edward Ayres \email{eayres@battelleecology.org}): The first issue is the sensor depths being incorrectly reported in the sensor_positions file. Our current data processing system does not allow a fix to this issue, so in the meantime we’ve added a file containing the installation depths and an associated readme file (both attached) to the data product documentation. This is mentioned in the Description section of the webpage you linked and you can find the files in the Documentation section. If you adapt your code to pull the depths from this file it should solve this issue.

# The second issue is that most of the soil moisture data were calculated using the manufacturers default calibration, rather than a soil-specific calibration, which causes the final quality flag to be raised. The fix to this issue is more complex depending on the option that you prefer. This issue is mentioned in the middle of the Abstract section of the webpage, which also includes a potential solution:

# The final quality flag for soil water content data is raised (i.e., VSWCFinalQF = 1) if it was calculated using the sensor manufacturer’s default calibration rather than a soil-specific calibration. However, for many use cases the default calibration is sufficient. To identify data that were solely flagged because they were generated using the default calibration, download the “Expanded” data package and identify rows where VSWCFinalQF = 1, VSWCAlphaQM < 10, VSWCBetaQM < 20, (VSWCFinalQFSciRvw = 0 or NA), and tempFailQM = 0.

# (Added 07-17-2023) Prior to RELEASE-2023, the final quality flag for soil water content data was raised (i.e., VSWCFinalQF = 1) when soil water content was calculated using the sensor manufacturer’s default calibration (calDefaultQM > 0) rather than a soil-specific calibration (calNEONQM > 0) . However, for many use cases the default calibration is sufficient. To identify data that were solely flagged because they were generated using the default calibration, download the “Expanded” data download package and identify rows where VSWCFinalQF = 1, VSWCAlphaQM < 10, VSWCBetaQM < 20, (VSWCFinalQFSciRvw = 0 or NA), and tempFailQM = 0. As of RELEASE-2023 (issued in January 2023), the final quality flag is no longer raised when the manufacturer's default calibration is used. The quality metrics indicating which calibration was used (calDefaultQM and calNEONQM) continue to be published in the expanded download package as informational metrics. (from the data documentation page for this product) - in this case we commented out the correction.
}
\examples{
\donttest{
# Download the soil water content data:
site_swc <- neonUtilities::loadByProduct(
#' dpID="DP1.00094.001",
site="SJER",
startdate="2020-05",
enddate="2020-05",
timeIndex = "30",
package="expanded",
check.size = F,
include.provisional = TRUE
)

Then correct the swc:
site_swc <- swc_correct(site_swc,site_name,download_date)
}
}
\author{
John Zobitz \email{zobitz@augsburg.edu}
}
